# Optimize getting the latest events in each room