As mentioned internally, this significantly speeds up getting the latest event in each room provided.

Output of `EXPLAIN ANALYSE` 
```
 Planning Time: 0.408 ms
 Execution Time: 9158.038 ms
```
for the original with a cold cache

```
 Planning Time: 0.296 ms
 Execution Time: 101.383 ms
```
with a warm cache.

The optimized query with a cold cache:
```
 Planning Time: 1.902 ms
 Execution Time: 3.752 ms
```
and with a warm cache:
```
 Planning Time: 0.425 ms
 Execution Time: 0.131 ms
```